// src/modules/preaching/interfaces/preaching.ts
export interface PreachingState {
    isPreachingDeleting: boolean;
    isPreachingLoading: boolean;
    isPreachingsLoading: boolean;
    preachings: Preaching[];
    selectedDate: Date;
    seletedPreaching: Preaching;
}
// ...
export interface Preaching {
    id: string;
    userId: string;
    day: string;
    initHour: string;
    finalHour: string;
    createdAt: string;
    updatedAt: string;
}
// ...
export type PreachingPayload = {
    preaching: Preaching
}

/**
 * SetSelectedDatePayload is an object with a property called selectedDate that is a Date.
 *
 * @property {Date} selectedDate - Date;
 */
export type SetSelectedDatePayload = {
    selectedDate: Date;
}

/**
 * SetPreachingsPayload is an object with a preachings property that is an array of Preaching objects.
 *
 * @property {Preaching[]} preachings - Preaching[]
 */
export type SetPreachingsPayload = {
    preachings: Preaching[];
}